Fixed MIMIC-III, MIMIC-IV using FP32 instead of FP64

+++ src/tsdm/datasets/mimic/mimic_iii_schema.py
@@ -8,7 +8,7 @@ from typing import Literal, TypeAlias
 import pyarrow as pa
 
 ID_TYPE = pa.uint32()
-VALUE_TYPE = pa.float32()
+VALUE_TYPE = pa.float64()
 TIME_TYPE = pa.timestamp("s")
 DATE_TYPE = pa.date32()
 BOOL_TYPE = pa.bool_()
